Fix a range of test paths, remove redundant test files
