+++ src/tests/sagas/simulator/runSaga.test.js
@@ -6,11 +6,11 @@ import {
   runSaga,
   getCoreOptionsFromState,
   initialiseCore
-} from '../../../sagas/simulatorSagas'
+} from '../../../features/simulator/sagas'
 
 import {
   RUN
-} from '../../../actions/simulatorActions'
+} from '../../../features/simulator/actions'
 
 
 describe('when testing the run saga', () => {
